django_ui/requirements.txt
# Core Django and web framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
celery>=5.3.4
redis>=5.0.1

# Database
psycopg2-binary>=2.9.9

# Authentication and security
djangorestframework-simplejwt>=5.3.0
django-oauth-toolkit>=2.2.0

# HTTP client and async
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.2

# Data processing and validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# AI and ML
openai>=1.3.6
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
nltk>=3.8.1
transformers>=4.35.2
torch>=2.1.1
accelerate>=0.25.0
# sentence-transformers>=2.2.2  # Replaced with Llama 3.2 3B
spacy>=3.7.2

# Academic and research APIs
arxiv>=1.4.8
scholarly>=1.7.11
crossref-commons>=0.0.7
pybtex>=0.24.0

# Audio and video processing
pydub>=0.25.1
moviepy>=1.0.3
openai-whisper>=20231117

# Document processing
PyPDF2>=3.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Translation and language
googletrans>=4.0.0rc1
langdetect>=1.0.9

# Email and notifications
django-mail-templated>=2.6.5

# Monitoring and logging
structlog>=23.2.0

# Development and testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Utils
python-dateutil>=2.8.2
pytz>=2023.3
Pillow>=10.1.0
markdown>=3.5.1
bleach>=6.1.0